core: Narrow the types passed to certain backend functions
Backend functions dealing with interfaces and alternate settings should
use a type whose range represents that of valid values for interfaces
and alternate settings. Switch to use uint8_t instead of int so that
backends do not have to cast values or do range checks.

+++ libusb/libusbi.h
@@ -942,7 +942,7 @@ struct usbi_os_backend {
 	 *   blocking
 	 * - another LIBUSB_ERROR code on other failure.
 	 */
-	int (*get_configuration)(struct libusb_device_handle *dev_handle, int *config);
+	int (*get_configuration)(struct libusb_device_handle *dev_handle, uint8_t *config);
 
 	/* Set the active configuration for a device.
 	 *
@@ -978,7 +978,7 @@ struct usbi_os_backend {
 	 *   was opened
 	 * - another LIBUSB_ERROR code on other failure
 	 */
-	int (*claim_interface)(struct libusb_device_handle *dev_handle, int interface_number);
+	int (*claim_interface)(struct libusb_device_handle *dev_handle, uint8_t interface_number);
 
 	/* Release a previously claimed interface.
 	 *
@@ -995,7 +995,7 @@ struct usbi_os_backend {
 	 *   was opened
 	 * - another LIBUSB_ERROR code on other failure
 	 */
-	int (*release_interface)(struct libusb_device_handle *dev_handle, int interface_number);
+	int (*release_interface)(struct libusb_device_handle *dev_handle, uint8_t interface_number);
 
 	/* Set the alternate setting for an interface.
 	 *
@@ -1012,7 +1012,7 @@ struct usbi_os_backend {
 	 * - another LIBUSB_ERROR code on other failure
 	 */
 	int (*set_interface_altsetting)(struct libusb_device_handle *dev_handle,
-		int interface_number, int altsetting);
+		uint8_t interface_number, uint8_t altsetting);
 
 	/* Clear a halt/stall condition on an endpoint.
 	 *
@@ -1076,7 +1076,7 @@ struct usbi_os_backend {
 	 * - another LIBUSB_ERROR code on other failure
 	 */
 	int (*kernel_driver_active)(struct libusb_device_handle *dev_handle,
-		int interface_number);
+		uint8_t interface_number);
 
 	/* Detach a kernel driver from an interface. Optional.
 	 *
@@ -1092,7 +1092,7 @@ struct usbi_os_backend {
 	 * - another LIBUSB_ERROR code on other failure
 	 */
 	int (*detach_kernel_driver)(struct libusb_device_handle *dev_handle,
-		int interface_number);
+		uint8_t interface_number);
 
 	/* Attach a kernel driver to an interface. Optional.
 	 *
@@ -1109,7 +1109,7 @@ struct usbi_os_backend {
 	 * - another LIBUSB_ERROR code on other failure
 	 */
 	int (*attach_kernel_driver)(struct libusb_device_handle *dev_handle,
-		int interface_number);
+		uint8_t interface_number);
 
 	/* Destroy a device. Optional.
 	 *
